Split workload module
